cancel & delete item

+++ app/src/main/java/com/vishnu/remindme/model/Reminder.kt
@@ -9,7 +9,7 @@ import kotlinx.parcelize.Parcelize
 @Parcelize
 data class Reminder(
     @PrimaryKey(autoGenerate = true)
-    var _id: Int = 0,
+    var _id: Long = 0,
     val title: String,
     val description: String?,
     val dueDate: Long,
